requirements.txt
streamlit>=1.45.0
requests>=2.32.3
numpy>=2.2.5
openai>=1.78.1
pillow>=11.2.1
qdrant-client>=1.14.2
python-dotenv
pandas
gradio-client